requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pymupdf
cohere
requests

langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-cohere
python-multipart

# MedCPT (dense retrieval) + ingestion
torch
transformers
trafilatura
beautifulsoup4
lxml
